Switched from rethinkdb to classical SQL-ORM
Since thinky actually just simulated default SQL bahvior anyway,
was very slow (At least on spinning disks, where 2-3 simple inserts
could take longer than 5s!) it was replaced with a more classical
approach using sequelize as an ORM for SQL-backends (For testing
sqlite, but could easily be switched to MySQL, Maria or Postgre).

Might be worse for big scaling, but it's unlikely that this project
will be used for anythin trello-sized.