Update to latest npm packages and grande fixup
To have the latest and greatest package version all were updated to the
current latests versions. This resulted in many tests failing due to
changes in sequelize behavior over major release versions. This was
fixed in now also looks much more like it was supposed to be than the
previous version which had some parts that felt like workarounds.