(Backend) Board user types changed and labels added
The board now does not split users into “owner” and “participants”
anymore. Instead it holds a single users list and flags them as admins
or default users in a join-table. The “owner” thus now only becomes the
first admin of the board.

Besides that the labels were implemented, which can now be created for
boards and assigned to cards.