Basic composer and routes added and corrected card movement
A basic composer was added to add new cards and columns (Still needs
quite some rework)

Also corrected the handler for cards-movement so it stores new
positions as expected and also prevents fragmentation (In single
columns or also in moves between columns). Those move-actions are also
now transactions so we can ensure that a move is fully stored or
totally reverted.