Better state handling for composer element and basic styling
The composer now holds a lifecycle of
- Waiting for add
- Composing
- save in progress

which is also reflected by the board (Adding columns) and the columns
(Adding cards) that reset the internal state accordingly after saving.
Besides that the escape-key can now be used to cancel a composing
action.

Also added basic less-styles for the composer for columns and cards.