Session update on refresh and add user to board handlers
Added rudamentary update to the login-component so it checks if the
user currently has an active session (e.g. has refreshed to browser).
If that’s the case the user-details will be used as credentials and the
session is resumed.

Also added first attempt for user to board adding in sidebar which
allows for searching all users by name and adding them to the board.
This still has to be unified for assigning users to cards and of course
extract the styling from the element directly to its own custom less
style.